update cookie test to only test against test cookie jar
The test cookie jar implementation is incomplete, and the corpus data
matches its behavior, rather than the behavior of a real cookie jar. The
test was previously attempting to optionally test against
HTTP::CookieJar, but due to a bug in the test just tested against the
test cookie jar a second time.

For now, change the test to only test against the test cookie jar. Also
never skip the test cookie jar test (such as for a syntax error). Fix
the test code to work against the specified cookie jar, even though we
are no longer trying to test any others. Even with incorrect data, the
test is still checking HTTP::Tiny's interaction with the cookie jar.

In the future, it would be good to update the test cookie jar to more
closely match the real behavior, fix the corpus data, and re-enable
testing against HTTP::CookieJar.