Implement keep_alive_timeout option
There's a well known race condition with http keep alive connections.
If the server closes the connection after the client has checked
the socket is still connected, but before it has prepared the request
to send it, then the client will receive an EPIPE error and fail
the entire request when it tries to write the request to the socket.

This means clients always have to check for this case and decide
what to do, which could be resend the request depending on when
exactly the EPIPE occured and whether the request was idempotent
or not.

Another work around for this is to make sure the client has a keepalive
reuse timeout that's less than the server keepalive time. So for instance
if the client has a 5 second timeout and the server has a 10 second
timeout, then if it's been > 5 seconds since the connection was last
re-used, the client will close it and open a new one, avoiding the case
where the server would close the connection on it.

This setup can be very useful when using HTTP::Tiny to send requests
to internal service endpoints and wanting as much as possible to
avoid this edge case failure more.